[feature] ftp file support save-mode 3
