+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/java/org/apache/seatunnel/e2e/connector/file/ftp/FtpFileIT.java
@@ -25,17 +25,27 @@ import org.apache.seatunnel.e2e.common.container.TestHelper;
 import org.apache.seatunnel.e2e.common.junit.DisabledOnContainer;
 import org.apache.seatunnel.e2e.common.util.ContainerUtil;
 
+import org.apache.commons.lang3.StringUtils;
+
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.TestTemplate;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.lifecycle.Startables;
+import org.testcontainers.shaded.com.github.dockerjava.core.command.ExecStartResultCallback;
 
+import com.github.dockerjava.api.command.ExecCreateCmdResponse;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 import java.util.stream.Stream;
 
 @DisabledOnContainer(
@@ -141,6 +151,65 @@ public class FtpFileIT extends TestSuiteBase implements TestResource {
         helper.execute("/orc/fake_to_ftp_file_orc.conf");
         // test write ftp root path excel file
         helper.execute("/excel/fake_source_to_ftp_root_path_excel.conf");
+        // test mult table and save_mode
+        testMultipleTableAndSaveMode(helper);
+    }
+
+    @SneakyThrows
+    private void testMultipleTableAndSaveMode(TestHelper helper) {
+        // test mult table and save_mode:RECREATE_SCHEMA DROP_DATA
+        String homePath = "/home/vsftpd/seatunnel";
+        String path1 = "/tmp/seatunnel_mult/text/source_1";
+        String path2 = "/tmp/seatunnel_mult/text/source_2";
+        Assertions.assertEquals(getFileListFromContainer(homePath + path1).size(), 0);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path2).size(), 0);
+        helper.execute("/text/multiple_table_fake_to_ftp_file_text.conf");
+        Assertions.assertEquals(getFileListFromContainer(homePath + path1).size(), 1);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path2).size(), 1);
+        helper.execute("/text/multiple_table_fake_to_ftp_file_text.conf");
+        Assertions.assertEquals(getFileListFromContainer(homePath + path1).size(), 1);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path2).size(), 1);
+        // test mult table and save_mode:CREATE_SCHEMA_WHEN_NOT_EXIST APPEND_DATA
+        String path3 = "/tmp/seatunnel_mult2/text/source_1";
+        String path4 = "/tmp/seatunnel_mult2/text/source_2";
+        Assertions.assertEquals(getFileListFromContainer(homePath + path3).size(), 0);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path4).size(), 0);
+        helper.execute("/text/multiple_table_fake_to_ftp_file_text_2.conf");
+        Assertions.assertEquals(getFileListFromContainer(homePath + path3).size(), 1);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path4).size(), 1);
+        helper.execute("/text/multiple_table_fake_to_ftp_file_text_2.conf");
+        Assertions.assertEquals(getFileListFromContainer(homePath + path3).size(), 2);
+        Assertions.assertEquals(getFileListFromContainer(homePath + path4).size(), 2);
+    }
+
+    @SneakyThrows
+    private List<String> getFileListFromContainer(String path) {
+        String command = "ls -1 " + path;
+        ExecCreateCmdResponse execCreateCmdResponse =
+                dockerClient
+                        .execCreateCmd(ftpContainer.getContainerId())
+                        .withCmd("sh", "-c", command)
+                        .withAttachStdout(true)
+                        .withAttachStderr(true)
+                        .exec();
+
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        dockerClient
+                .execStartCmd(execCreateCmdResponse.getId())
+                .exec(new ExecStartResultCallback(outputStream, System.err))
+                .awaitCompletion();
+
+        String output = new String(outputStream.toByteArray(), StandardCharsets.UTF_8).trim();
+        List<String> fileList = new ArrayList<>();
+        log.info("container path file list is :{}", output);
+        String[] files = output.split("\n");
+        for (String file : files) {
+            if (StringUtils.isNotEmpty(file)) {
+                log.info("container path file name is :{}", file);
+                fileList.add(file);
+            }
+        }
+        return fileList;
     }
 
     @AfterAll
