+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-file-ftp-e2e/src/test/resources/text/multiple_table_fake_to_ftp_file_text_2.conf
@@ -89,7 +89,7 @@ sink {
     port = 21
     user = seatunnel
     password = pass
-    path = "/tmp/seatunnel_mult/text/${table_name}"
+    path = "/tmp/seatunnel_mult2/text/${table_name}"
     source_table_name = "ftp"
     row_delimiter = "\n"
     partition_dir_expression = "${k0}=${v0}"
@@ -99,7 +99,7 @@ sink {
     filename_time_format = "yyyy.MM.dd"
     is_enable_transaction = true
     compress_codec = "lzo"
-    "schema_save_mode"="RECREATE_SCHEMA"
-    "data_save_mode"="DROP_DATA"
+    "schema_save_mode"="CREATE_SCHEMA_WHEN_NOT_EXIST"
+    "data_save_mode"="APPEND_DATA"
   }
 }
